extract content of the pull request error code generation handler into functions
